backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
aiomysql>=0.1.1
pydantic>=2.0.0
passlib>=1.7.4
bcrypt>=4.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
